Marketing_Crew/Marketing_crew/pyproject.toml
[project]
name = "marketing-crew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.11.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "streamlit>=1.28.0",
]
